feat: Add maxPoints option to Leaflet Ruler control
This commit adds a new option, maxPoints, to the Leaflet Ruler control. The maxPoints option allows users to set a maximum number of points in one line when measuring distances on the map. When the maximum number of points is reached, the line is automatically closed.

The maxPoints option is added to both the README.md file and the leaflet-ruler.js file.

Refs: #4